feat: add ownership filter to delete supplier route

+++ src/supplier/supplier.service.ts
@@ -22,8 +22,8 @@ export class SupplierService
 		return await this.prisma.supplier.findUnique( { where: { id: id, client_id: user_id } } );
 	}
 
-	async delete_by_id( id: number )
+	async delete_by_id( user_id: number, id: number )
 	{
-		return await this.prisma.supplier.delete( { where: { id: id } } );
+		return await this.prisma.supplier.delete( { where: { id: id, client_id: user_id } } );
 	}
 }
